test: integration test for delete function. Change test execution to …

+++ src/service.rs
@@ -30,15 +30,14 @@ impl DbService {
     */
 
     pub fn destroy_db(&mut self) -> (bool, String) {
-        let (is_open, msg) =self.open_db();
+        let (is_open, msg) = self.open_db();
         if !is_open {
             return (false, msg);
-        } else {
-            match self.rocks_db_facade.destroy_db(DB_PATH) {
-                Ok(()) => return (true, String::from("Destroyed database at path '") + DB_PATH + "'"),
-                Err(e) => return (false, String::from("Error when trying to destroy database at path '")
-                    + DB_PATH + "': " + &e.to_string()),
-            }
+        } 
+        match self.rocks_db_facade.destroy_db(DB_PATH) {
+            Ok(()) => return (true, String::from("Destroyed database at path '") + DB_PATH + "'"),
+            Err(e) => return (false, String::from("Error when trying to destroy database at path '")
+                + DB_PATH + "': " + &e.to_string()),
         }
     }
 
@@ -46,12 +45,11 @@ impl DbService {
         let (is_open, msg) = self.open_db();
         if !is_open {
             return (false, msg);
-        } else {
-            match self.rocks_db_facade.write_db(key, value) {
-                Ok(()) => return (true, String::from("Wrote key '") + key + "' and value '" + value + "'"),
-                Err(e) => return (false, String::from("Error when trying to write key '") + key
-                    + "' and value '" + value + "': " + &e.to_string()),
-            }
+        } 
+        match self.rocks_db_facade.write_db(key, value) {
+            Ok(()) => return (true, String::from("Wrote key '") + key + "' and value '" + value + "'"),
+            Err(e) => return (false, String::from("Error when trying to write key '") + key
+                + "' and value '" + value + "': " + &e.to_string()),
         }
     }
 
@@ -59,12 +57,18 @@ impl DbService {
         let (is_open, msg) = self.open_db();
         if !is_open {
             return (false, msg,  String::from(""));
-        } else {
-            match self.rocks_db_facade.read_db(key) {
-                Ok(value) => return (true, String::from("Retrieved value '") + &value + "' from key '" + key + "'", value),
-                Err(e) => return (false, String::from("Error when trying to retrieve from key '") + key
-                    + "': " + &e.to_string(), String::from("")),
-            }
+        } 
+        match self.rocks_db_facade.read_db(key) {
+            Ok(value) => return (true, String::from("Retrieved value '") + &value + "' from key '" + key + "'", value),
+            Err(e) => return (false, String::from("Error when trying to retrieve from key '") + key
+                + "': " + &e.to_string(), String::from("")),
+        }
+    }
+
+    pub fn check_if_key_exists(&mut self, key:&str) -> bool {
+        match self.rocks_db_facade.read_db(key) {
+            Ok(_value) => return true,
+            Err(_e) => return false,
         }
     }
 
@@ -72,12 +76,16 @@ impl DbService {
         let (is_open, msg) = self.open_db();
         if !is_open {
             return (false, msg);
-        } else {
+        } 
+
+        if self.check_if_key_exists(key) {
             match self.rocks_db_facade.delete_db(key) {
                 Ok(()) => return (true, String::from("Deleted key '") + key + "'"),
                 Err(e) => return (false, String::from("Error when trying to delete key '") + key
                     + "': " + &e.to_string()),
             }
+        } else {
+            return (false, String::from("Key '") + key + "' does not exist!")
         }
     }
 }
